Add avatar names
