+++ app/src/main/java/io/ipoli/android/store/viewmodels/PetViewModel.java
@@ -5,7 +5,7 @@ import android.support.annotation.DrawableRes;
 
 import org.threeten.bp.LocalDate;
 
-import io.ipoli.android.store.PetAvatar;
+import io.ipoli.android.player.PetAvatar;
 
 /**
  * Created by Venelin Valkov <venelin@curiousily.com>
