+++ app/src/main/res/values/strings.xml
@@ -744,6 +744,18 @@
     <string name="most_popular">Most popular</string>
     <string name="jumbo_pack">Jumbo pack</string>
     <string name="jumbo_pack_desc">For the wealthy and famous</string>
+    <string name="avatar_name_ipoli_classic">Classic iPoli</string>
+    <string name="avatar_name_ipoli_yellow_glasses">Fancy iPoli</string>
+    <string name="avatar_name_macarena">Lady Luck</string>
+    <string name="avatar_name_macarena_mexican">Mamacita</string>
+    <string name="avatar_name_blondy">Miss Charming</string>
+    <string name="avatar_name_green_eyes">Angel Eyes</string>
+    <string name="avatar_name_pipilota">Butter Cup</string>
+    <string name="avatar_name_old_pirate">Treasure guy</string>
+    <string name="avatar_name_beard_guy">Mr. Right</string>
+    <string name="avatar_name_dwight">Muscle man</string>
+    <string name="avatar_name_michael">Bad boy</string>
+    <string name="avatar_name_tobby">Brainiac</string>
 
 
     <string-array name="player_titles">
